+++ trixnity-client/integration-tests/src/jvmTest/kotlin/net/folivo/trixnity/client/integrationtests/testUtils.kt
@@ -63,7 +63,6 @@ suspend fun MatrixClientServerApiClient.register(
     val registerResult = registerStep.authenticate(AuthenticationRequest.Dummy).getOrThrow()
     registerResult.shouldBeInstanceOf<UIA.Success<Register.Response>>()
     val (userId, createdDeviceId, accessToken, _, refreshToken) = registerResult.value
-    println("--- " + registerResult.value.accessTokenExpiresInMs)
     requireNotNull(createdDeviceId)
     requireNotNull(accessToken)
     return Result.success(MatrixClient.LoginInfo(userId, createdDeviceId, accessToken, refreshToken))